Move NDTensors HDF5 support to package extension
